requirements.txt
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.1
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langchainhub==0.1.16
openai==1.30.3
tiktoken==0.7.0
python-dotenv
Django==4.2.9
djangorestframework==3.15.1
django-cors-headers==4.3.1
pymupdf
faiss-cpu